+++ src/arm/linux/aarch32-isa.c
@@ -23,7 +23,7 @@ void cpuinfo_set_wcid(uint32_t wcid) {
 #endif
 
 void cpuinfo_arm_linux_decode_isa_from_proc_cpuinfo(
-	uint32_t features,
+	uint64_t features,
 	uint64_t features2,
 	uint32_t midr,
 	uint32_t architecture_version,
